Maintain a user's information upon app closure
Before, a user could log in and I was saving their username and
password, but things like their UserId, e-mail, etc. were lost. In other
words, I was not maintaining a User object when then application closed.
I added a dictionary with key = "User" and value = the JSON
representation of the logged in user. Now whenever the user logs in, I
serialize that object and save it in the AppSettings. When the
application is opened, I look in the AppSettings for that object and if
it's there, I restore previously logged in user as App.LoggedInUser.

When a user logs out, I remove the User dictionary from AppSettings
(just like I do with Username and Password).