Extracting BitMapImage using GetThumbnail method
This is actually a pretty big update and frankly the project is kind of
in shambles. The biggest thing in this commit is working on extracting
an image from a video, then binding it to the UI. This task wasn't
easily accomplished and there was some experimentation code involved and
left in the project. Some of it is commented out - some of it is not.
There is much to do before the final presentations.