Lots of new code, see description
Worked on video feed. There were some threading issues when getting
thumbnail, uploading to SkyDrive, etc. so I toyed around with that.

I added a button on the GroupedItemsPage to sign up for an account and
also a page for signing up. This way when people are demoing our app we
can allow them to create their Archive account there. I'd like to have
more robust accounts including and image and links to other sites (i.e.,
Twitter, e-mail, etc.)